refactor locationsearch to component
