+++ src/pages/signup.js
@@ -12,18 +12,11 @@ import {
   TextArea,
   TextInput,
 } from 'grommet'
-import algoliasearch from 'algoliasearch/lite'
-import { InstantSearch, connectAutoComplete } from 'react-instantsearch-dom'
-import AsyncSelect from 'react-select/async'
 
 import Page from '../components/shared/Page'
 import Spinner from '../components/shared/Spinner'
+import LocationSearch from '../components/signup/LocationSearch'
 import { businessOptions, offeringOptions } from '../components/signup/presets'
-
-const searchClient = algoliasearch(
-  'TGPZX7CMYY',
-  '859c34030d228a6188c83731bb6e456f'
-)
 
 const API_HOST =
   process.env.NODE_ENV === 'production'
@@ -131,49 +124,6 @@ const TextAreaField = ({ errorMsg, label, ...rest }) => {
       <TextArea {...rest} onFocus={onFocus} onBlur={onBlur} />
       {errorMsg && <p>errorMsg</p>}
     </TextInputContainer>
-  )
-}
-
-const Autocomplete = ({
-  onChange,
-  currentOption,
-  hits,
-  currentRefinement,
-  refine,
-}) => {
-  const loadOptions = (_, callback) => {
-    callback(hits)
-  }
-  const handleInputChange = input => {
-    if (input) {
-      refine(input)
-    }
-  }
-  const handleChoose = input => {
-    refine(input)
-    onChange(input)
-  }
-  return (
-    <AsyncSelect
-      value={currentOption}
-      defaultOptions
-      loadOptions={loadOptions}
-      onInputChange={handleInputChange}
-      onChange={handleChoose}
-      formatOptionLabel={option => (
-        <span>
-          {option.suburb} {option.state} {option.postcode}
-        </span>
-      )}
-    />
-  )
-}
-const LocationSearch = ({ onChange, currentOption }) => {
-  const CustomAutocomplete = connectAutoComplete(Autocomplete)
-  return (
-    <InstantSearch searchClient={searchClient} indexName="prod_suburb_centroid">
-      <CustomAutocomplete onChange={onChange} currentOption={currentOption} />
-    </InstantSearch>
   )
 }
